@@ -1,12 +1,12 @@
 plugins {
-	id 'org.springframework.boot' version '2.2.1.RELEASE'
-	id 'io.spring.dependency-management' version '1.0.8.RELEASE'
-	id 'java'
+	id 'org.springframework.boot' version '2.7.8'
+	id 'io.spring.dependency-management' version '1.0.15.RELEASE'
+	id 'java-library'
 }
 
 group = 'org.easeci'
 version = ''
-sourceCompatibility = '11'
+sourceCompatibility = '17'
 
 configurations {
 	compileOnly {
@@ -26,20 +26,25 @@ dependencies {
 	implementation('org.springframework.boot:spring-boot-starter-jdbc')
 	implementation('org.springframework.boot:spring-boot-starter-security')
 	implementation('org.springframework.boot:spring-boot-starter-mail')
-    implementation 'org.projectlombok:lombok:1.18.18'
-    compile group: 'io.vavr', name: 'vavr', version: '0.10.3'
-	compile group: 'org.thymeleaf.extras', name: 'thymeleaf-extras-springsecurity5', version: '3.0.4.RELEASE'
-	compile group: 'org.postgresql', name: 'postgresql', version: '42.2.8'
-	compileOnly 'org.projectlombok:lombok'
-	annotationProcessor 'org.projectlombok:lombok'
+	implementation 'jakarta.validation:jakarta.validation-api:3.0.2'
+	implementation 'javax.validation:validation-api:2.0.1.Final'
+
+	implementation 'org.hibernate.validator:hibernate-validator:8.0.0.Final'
+	implementation 'org.projectlombok:lombok:1.18.26'
+	implementation group: 'io.vavr', name: 'vavr', version: '0.10.3'
+	implementation group: 'org.thymeleaf.extras', name: 'thymeleaf-extras-springsecurity5', version: '3.0.4.RELEASE'
+	implementation group: 'org.postgresql', name: 'postgresql', version: '42.2.8'
+	compileOnly 'org.projectlombok:lombok:1.18.26'
+	annotationProcessor 'org.projectlombok:lombok:1.18.26'
 	testImplementation('org.springframework.boot:spring-boot-starter-test') {
 		exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
 	}
 	testImplementation 'io.projectreactor:reactor-test'
-	testCompile group: 'com.h2database', name: 'h2', version: '1.4.200'
+	testImplementation group: 'com.h2database', name: 'h2', version: '1.4.200'
 
-	compile group: 'org.webjars.npm', name: 'ace-builds', version: '1.4.11'
-	compile group: 'org.webjars', name: 'bootstrap', version: '4.3.1'
+	implementation group: 'org.webjars.npm', name: 'ace-builds', version: '1.14.0'
+	implementation 'org.webjars:bootstrap:5.2.3'
+
 }
 
 test {
