@@ -5,9 +5,9 @@ import lombok.NoArgsConstructor;
 import org.easeci.registry.domain.user.RegistryUser;
 import org.hibernate.validator.constraints.Length;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Pattern;
 
 @Data
 @NoArgsConstructor
